% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAVHRRannualts.R
\name{addAVHRRannualts}
\alias{addAVHRRannualts}
\title{Add terrestrial AVHRR time series to limnological time series}
\usage{
addAVHRRannualts(lakelist, avhrrdata, buffdist = 5000, aggfun = "mean")
}
\arguments{
\item{lakelist}{a list containing \code{lakeinfo} and \code{lakedata}, as output from \code{makeLAGOSannualts}.}

\item{avhrrdata}{a raster stack of annual AVHRR metrics}

\item{buffdist}{the distance (in meters) of the radius around each lake to draw terrestrial productivity information from.}

\item{aggfun}{a string indicating how to aggregate values from multiple satellite measurements within the buffer. Defaults to "mean",
"median" is also available.}
}
\value{
\code{addAVHRRannualts} returns an object of class \code{list}. Slots are:
\item{lakeinfo}{A data frame of lake information.}
\item{lakedata}{A list of data frames containing the lake data. The avhrr data are appended here.}
}
\description{
\code{addAVHRRannualts} extracts terrestrial AVHRR time series and appends them to limnological time series produced by,
e.g., \code{makeLAGOSannualts}.
}
\details{
Nope.
}
\examples{
#need to add some

}
\seealso{
\code{\link{addAVHRRnnualts}}, \code{\link{makeIowaALMannualts}}
}
\author{
Jonathan Walter, \email{jaw3es@virginia.edu}
}
