% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPRISMts.R
\name{addPRISMts}
\alias{addPRISMts}
\title{Add terrestrial AVHRR time series to limnological time series}
\usage{
addPRISMts(lakelist, prismdata, var)
}
\arguments{
\item{lakelist}{a list containing \code{lakeinfo} and \code{lakedata}, as output from \code{makeLAGOSannualts}.}

\item{prismdata}{a raster stack of annual AVHRR metrics}

\item{var}{what variable is being handled? currently, ppt and tmean are accepted.}
}
\value{
\code{addPRISMts} returns an object of class \code{list}. Slots are:
\item{lakeinfo}{A data frame of lake information.}
\item{lakedata}{A list of data frames containing the lake data. The PRISM data are appended here.}
}
\description{
\code{addPRISMts} extracts PRISM time series and appends them to limnological time series produced by,
e.g., \code{makeLAGOSannualts}.
}
\details{
Extracts the PRISM data using latitude-longitude coordinates provided by LAGOS
}
\examples{
#need to add some

}
\seealso{
\code{\link{addAVHRRannualts}}, \code{\link{makeIowaALMannualts}}
}
\author{
Jonathan Walter, \email{jaw3es@virginia.edu}
}
