% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBufferDist.R
\name{calcBufferDist}
\alias{calcBufferDist}
\title{Create annualized time series from the LAGOSNE database}
\usage{
calcBufferDist(sa, ltype, minbuff = 2500)
}
\arguments{
\item{sa}{a numeric vector of lake surface areas (hectares)}

\item{ltype}{lake type; see details}

\item{minbuff}{the minimum buffer radius (meters)}
}
\value{
\code{calcBufferDist} returns a numeric vector of buffer distances (meters)
}
\description{
\code{calcBufferDist} calculates a buffer distance based on lake surface area
}
\details{
Computed buffer distances are based on a log-linear relationship between lake surface area and watershed area,
parameterized on empirical data (Walter, et al., unpublished data) with a further simplifying assumption that lakes are circular. 
Setting a minimum buffer radius is useful for cases when lakes are small relative to the grain size of satellite data.
}
\examples{
#need to add some

}
\seealso{
\code{\link{makeIowaALMannualts},\link{addAVHRRannualts}}
}
\author{
Jonathan Walter, \email{jaw3es@virginia.edu}
}
