% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanAnnualts.R
\name{cleanAnnualts}
\alias{cleanAnnualts}
\title{Clean annual time series of lake and terrestrial NDVI data}
\usage{
cleanAnnualts(
  indat,
  ymin = 20,
  maxNA = 2,
  optim = "chla",
  timespan = NULL,
  fill.method = "median"
)
}
\arguments{
\item{indat}{an input data list, e.g. output from 'makeLAGOSannualts'}

\item{ymin}{minumum time series length}

\item{maxNA}{maximum number of missing observations per time series. If 0< maxNA<1, maxNA is the maximum proportion of NAs allowed.}

\item{timespan}{a vector giving the min and max of a range of years which data will be restricted to.}

\item{fill.method}{the method for filling missing observations. Currently, only 'median' is implemented.}
}
\value{
\code{cleanAnnualts} returns an object of class \code{list}. Slots are:
\item{lakeinfo}{A data frame of lake information generated by, e.g., \code{makeLAGOSannualts}.}
\item{lakedata}{A list of data frames containing the lake data, as generated by, e.g., \code{makeLAGOSannaults}.}
}
\description{
\code{cleanAnnualts} performs routine cleaning and filtering on annualized time series.
}
\details{
\code{cleanAnnualts} takes as input a list generated by \code{makeLAGOSannualts}, possibly augmented by \code{makeAVHRRannualts},
\code{makeIowaALMannualts}, or other dataset building functions. It cleans the data, finding optimal stretches of data meeting length
and completeness requirements set by \code{ymin} and \code{maxNA}. Currently, \code{optim} sets a single variable to optimize time series
based on, but possibly future updates will include opimitization on multiple variables. Missing values are filled only if they are non-consecutive. 
If no variables meet requirements, the whole lake is dropped from the output. If some, but not all variables meet requirements, 
data are returned with the offending variables ommitted. Data should still be inspected and explored prior to 
analysis, and in some cases some manual bespoke cleaning may be necessary for certain lakes and variables.
}
\examples{
#need to add some

}
\seealso{
\code{\link{makeLAGOSannualts}}
}
\author{
Jonathan Walter, \email{jaw3es@virginia.edu}
}
