% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNamestoLAGOS.R
\name{fixNamestoLAGOS}
\alias{fixNamestoLAGOS}
\title{Fix variable names to match LAGOS conventions}
\usage{
fixNamestoLAGOS(datalist, origin)
}
\arguments{
\item{datalist}{a data list, as output from \code{makeIowaALMannualts} or \code{makeMNMPCAannualts}.}

\item{origin}{a character string indicating the origin of the data. See details for options.}
}
\value{
\code{fixNamestoLAGOS} returns an object of class \code{list}. Slots are:
\item{lakeinfo}{A data frame of lake information.}
\item{lakedata}{A list of data frames containing the lake data.}
}
\description{
\code{fixNamestoLAGOS} takes variables from disparate datasets and reconciles their names to match LAGOS conventions
}
\details{
Options of \code{origin} include "Iowa ALM", "Minnesota MPCA", and "Wisconsin DNR".
}
\examples{
#need to add some

}
\seealso{
\code{\link{makeLAGOSannualts}}
}
\author{
Jonathan Walter, \email{jaw3es@virginia.edu}
}
