% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIowaALMannualts.R
\name{makeIowaALMannualts}
\alias{makeIowaALMannualts}
\title{Create annualized time series from Iowa ALM data}
\usage{
makeIowaALMannualts(
  almdat,
  lakechar,
  getvars,
  aggfun = "gsmean",
  timespan = NULL,
  minmos = 3,
  minobs = 3
)
}
\arguments{
\item{almdat}{a data frame of Iowa Ambient Lakes Monitoring (ALM) data, typically 'Iowa Synchrony Lakes.csv' on Box}

\item{lakechar}{a lake characteristics data frame, typically 'Iowa Lake Characteristics.csv' on Box}

\item{aggfun}{a string indicating how to summarize multiple observations to an annual timescale. See details.}

\item{timespan}{a vector giving the min and max of a range of years which data will be restricted to.}

\item{minmos}{the minimum number of months required to have data from for an annual value to be computed.}

\item{minobs}{the minimum number of observations, spread over \code{>= minmos}, for an annual value to be computed}

\item{tsvars}{a character vector of variables to extract and construct time series from.}

\item{infovars}{a character vector of lake characteristic variables to include in \code{lakeinfo}. See details.}
}
\value{
\code{makeIowaALMannualts} returns an object of class \code{list}. Slots are:
\item{lakeinfo}{A data frame of lake information.}
\item{lakedata}{A list of data frames containing the lake data.}
}
\description{
\code{makeIowaALMannualts} formats Iowa ALM data to match output from \code{makeLAGOSannualts}.
}
\details{
Currently, the only accepted value of \code{aggfun} is \code{"gsmean"}, corresponding to the growing season mean.
We consider the growing season to last May to September. Output \code{lakeinfo} is a data frame containing the numeric id code, name, 
and geographic coordinates of lakes, plus any further variables listed in \code{infovars}. Each element of \code{lakedata} corresponds to a lake
in \code{lakes} and is a data frame in which columns are variables in \code{tsvars} and rows correspond to years.
}
\examples{
#need to add some

}
\seealso{
\code{\link{makeLAGOSannualts}}
}
\author{
Jonathan Walter, \email{jaw3es@virginia.edu}
}
